Temporarily rm to fix git-subtree error.
The error is this one:
`Can't squash-merge: 'depends/ChibiOS-RPi' was never added.`

Apparently the earlier directory move wasn't the right way to do this.
More info here:
https://stackoverflow.com/questions/9777564/git-subtree-pull-complications

The subtree command is probably looking for 'depends/ChibiOS-RPi' in the
original subtree commit, but it is finding just 'ChibiOS-RPi' instead,
which it is ignoring, so it never really finds anything and thus
assumes that the prefix was never added. Darn.

So the plan is to remove it and add it back in, thus generating a commit
with the correct prefix metadata.