Got PHT readings from the MS8607. Good!
So I got the pressure-humidity-temperature sensor working, but the code
is a mess: printf debugging everywhere and the MS8607 driver integration
is, uh, minimal. I might want to clean up error handling a lot, too, and
that involves changing a bunch of driver code+API, so that's not done
here, but at least I figured out that there is something up with
hardware float and chprintf on this processor+toolchain+chibios and got
the sensor to work.