Actual main.c for latest version of ms8607 driver.
Previous version did some experimenting with the humidity sensor user
register and had some commented-out code for taking readings, and that
code was experimental code for using a non-blocking version of the
ms8607 driver, a version that I can't finish right now.

So I have switched the back over to taking pressure+temp+humidity
readings and fixed it up to work with the most recent version of the
ms8607 driver, including the all-important no-more-global-state
upgrade.

(I'm giving up on the non-blocking version of the ms8607 driver
for now; I can't do it in my current circumstances, as much as I'd like
to. For what I'm doing, the blocking logic will be sufficient.)