Move pressure set-point for testing purposes.
It's at like 400mbar right now. Not for human use. Instead,
this setting is to see how much pressure the piezo pump can
deliver, and it will probably be left on for a while, because
it is also useful for leak-testing the prototype as I add and
remove fittings and components.