Bugfix: filtered pressure values no longer off by 1/4.
To smooth out any noise in pressure readings, a small number of
preceding samples are used in a weighted average (biased towards
recency). The bug was that the less-recent samples were not being
counted because of an index-offsetting error in the for-loop that
accumulates sample sums. This resulted in seeing pressure values
that were 3/4 of what they should be. So that's fixed, and all
of the readings make sense now.