Use newer version of PHT drivers.
This involves deduplicating the code for ChibiOS<->TEPHT
integration, as the previous API made it difficult to
share the `*_host_functions` structure (and the functions
that it points to) between different PHT sensor drivers.
Those drivers now share a common definition for these things,
so we can remove the effectively-duplicated definitions
and have code that is closer to complying with the DRY principle.