Add 3rd sensor, make sensors indexable.
The 3rd sensor is a "return" sensor that measures the pressure
in a chamber after the air returns from the ears, back to the
machine. This sensors can potentially help with leak detection,
flow tuning, and provide a more accurate measurement of the
pressure actually being provided at the ear.

This commit also changes how the I2C multiplexer is used:
Before the port selection bits would only be sent to the
multiplexer when they change.
Now they are sent EVERY time this code communicates with one
of the sensors.
This change was required to overcome a problem where the
system would receive NACK signals from sensors after a very
short amount of communication, and after that, would never
respond with anything besides NACK, which very quickly
made the whole device fail to function. Perhaps the multiplexer
was closing the channels or being reset, because simply
re-issuing the port selection bits on every I2C communication
was enough to resolve this issue.