Experimental configuration: timing tests.
This commit turns off the pressure sensor code and instead
compares a few different ways to measure time using the BCM2835.

The {CLO,CHI} registers seem to give the best measurement
of time-since-start, and were found to also have microsecond
granularity, which is great. This determines how timing
will be done for the rest of the project (or, at least,
the Raspberry Pi Zero version).